requirements.txt
streamlit
torch
transformers
pillow
openai
python-dotenv
safetensors
opencv-python
matplotlib
numpy
timm
grad-cam